inputado o Get e Post do Usuario, Login e Autenticador de idade

+++ Data/Dtos/BuscaUsuarioDto.cs
@@ -3,7 +3,7 @@
     public class BuscaUsuarioDto
     {
         public string Nome { get; set; }
-        public int CPF { get; set; }
+        public long CPF { get; set; }
         public string UF { get; set; }
         public DateTime DataNascimento { get; set; }
     }
